safety

+++ Core/Src/battery/battery.h
@@ -1,7 +1,7 @@
 /* Core/Inc/battery.h
  *
  * Exposes:
- *  • ReadBatteryVoltage()  — raw voltage in volts, or –1.0f on error
+ *  • ReadBatteryVoltage()  — raw battery voltage in mV, or negative error code
  *  • GetBatteryLevel()     — voltage + percentage, returns status
  */
 
@@ -16,23 +16,23 @@ extern "C" {
 
 /* Status codes for GetBatteryLevel() */
 typedef enum {
-    BATTERY_OK       = 0,
+    BATTERY_OK        =  0,
     BATTERY_ADC_ERROR = -1,
     BATTERY_TIMEOUT   = -2,
     BATTERY_ERROR     = -3
 } Battery_Status_t;
 
 /**
- * @brief  Read the battery voltage (halved by divider + buffered).
- * @return VBAT in volts, or –1.0f on error.
+ * @brief  Read the battery voltage (halved by divider).
+ * @return Battery voltage in millivolts [mV], or negative error code.
  */
-float ReadBatteryVoltage(void);
+int32_t ReadBatteryVoltage(void);
 
 /**
  * @brief  Read VBAT and compute a 0–100% percentage.
- * @param  out_v     pointer to float → battery voltage [V]
+ * @param  out_v     pointer to uint32_t → battery voltage [mV]
  * @param  out_pct   pointer to uint8_t → battery % (0…100)
- * @return BATTERY_OK or BATTERY_ERROR
+ * @return BATTERY_OK or a negative error code.
  */
 Battery_Status_t GetBatteryLevel(uint32_t *out_v, uint8_t *out_pct);
